refactor: 💡 split bot code into separate files under package
